+++ vitest.config.ts
@@ -2,9 +2,23 @@ import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
   test: {
-    include: ['test/**/*.spec.ts'],
-    environment: 'node',
     pool: 'forks',
-    environmentMatchGlobs: [['test/client*.spec.ts', 'jsdom']],
+    projects: [
+      {
+        test: {
+          name: 'node',
+          environment: 'node',
+          include: ['test/**/*.spec.ts'],
+          exclude: ['test/client*.spec.ts'],
+        },
+      },
+      {
+        test: {
+          name: 'client',
+          environment: 'jsdom',
+          include: ['test/client*.spec.ts'],
+        },
+      },
+    ],
   },
 })